+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This file is used to list changes made in each version of the gem.
   - Remove version from compose.yml.
   - Auto-stop staging container after 7 days.
   - Use `rails` chart in Gitlab CI templates.
+  - Move entrypoints locations to `bin` to be rails-standard
 - Version Docker generator.
 - Add a generator for project initialization:
   - Add default configuration for production and staging environments.